Add proper python3 directory to the path
python3 is not in the path by default; instead you need to run
"module load conda" to get it. However, this code submits python
scripts as binary, so there is nowhere to run the "module load"
command. The quickest way around this is to include the full path
to python3 in the PATH environment variable.